Move torture tests into their own test program.
