+++ bfffs-core/tests/torture/mod.rs
@@ -1,0 +1,6 @@
+pub mod util {
+    include!("../functional/util.rs");
+}
+
+mod fs;
+use util::PoolBuilder;
